move task test to correct package
