+++ src/test/java/co/com/bancolombia/task/GenerateDrivenAdapterTaskTest.java
@@ -1,4 +1,4 @@
-package co.com.bancolombia;
+package co.com.bancolombia.task;
 
 import co.com.bancolombia.exceptions.CleanException;
 import co.com.bancolombia.task.GenerateDrivenAdapterTask;
