+++ src/test/java/co/com/bancolombia/task/GenerateStructureTaskTest.java
@@ -1,4 +1,4 @@
-package co.com.bancolombia;
+package co.com.bancolombia.task;
 
 import co.com.bancolombia.task.GenerateStructureTask;
 import org.gradle.api.Project;
